Do not restore the original number of workspaces
The extension is disabled when GNOME is inactive or locked and
the reset of workspaces caused windows to move from one workspace
to the other. Now changing the number of workspaces in this extension
will permanently change the global number of workspaces.

If workspaces were set to "dynamic" before the extension is enabled,
windows will still jump around.

References #29