remove unused depracated Tweener import
Tweener was removed and replaced with Clutter's implicit animations since gnome shell 3.37.1, and while 4830020695477e2658633009c5449c45affca9fe updated already depracated by then animation code, it left out the import itself.

This results in incompatibility with gnome 3.37.90 and is not used anywhere, so remove it from the code for good.